requirements.txt
asgiref==3.6.0
Django==4.1.5
django-material-admin==1.8.6
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pydantic==1.10.4
python-decouple==3.7
python-dotenv==0.21.0
pytz==2022.7
qrcode==7.3.1
segno==1.5.2
sqlparse==0.4.3
typing_extensions==4.4.0
